+++ src/Community.PowerToys.Run.Plugin.CursorWorkspaces/obj/ARM64/Release/Community.PowerToys.Run.Plugin.CursorWorkspaces.AssemblyInfo.cs
@@ -13,10 +13,11 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("Community.PowerToys.Run.Plugin.CursorWorkspaces")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Release")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+be0e3c2d51f1dc1e5bc16114267920afda046938")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+9c0847c8637cc693abbe3dc2736f90196a834943")]
 [assembly: System.Reflection.AssemblyProductAttribute("Community.PowerToys.Run.Plugin.CursorWorkspaces")]
 [assembly: System.Reflection.AssemblyTitleAttribute("Community.PowerToys.Run.Plugin.CursorWorkspaces")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]
+[assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Community.PowerToys.Run.Plugin.CursorWorkspaces.Tests")]
 [assembly: System.Runtime.Versioning.TargetPlatformAttribute("Windows7.0")]
 [assembly: System.Runtime.Versioning.SupportedOSPlatformAttribute("Windows7.0")]